+++ CHANGELOG.md
@@ -31,3 +31,5 @@
 
 
 ### Internal changes
+
+- Fix tests that assert on absolute paths; must assert on paths relative to the project folder.